+++ emr-20210320/alibabacloud_emr20210320/client.py
@@ -286,6 +286,110 @@ class Client(OpenApiClient):
         runtime = util_models.RuntimeOptions()
         return await self.create_node_group_with_options_async(request, runtime)
 
+    def create_report_with_options(
+        self,
+        request: emr_20210320_models.CreateReportRequest,
+        runtime: util_models.RuntimeOptions,
+    ) -> emr_20210320_models.CreateReportResponse:
+        """
+        Currently we only support taihao platform
+        
+        @param request: CreateReportRequest
+        @param runtime: runtime options for this request RuntimeOptions
+        @return: CreateReportResponse
+        """
+        UtilClient.validate_model(request)
+        query = {}
+        if not UtilClient.is_unset(request.cluster_id):
+            query['ClusterId'] = request.cluster_id
+        if not UtilClient.is_unset(request.region_id):
+            query['RegionId'] = request.region_id
+        if not UtilClient.is_unset(request.select_timestamp):
+            query['SelectTimestamp'] = request.select_timestamp
+        req = open_api_models.OpenApiRequest(
+            query=OpenApiUtilClient.query(query)
+        )
+        params = open_api_models.Params(
+            action='CreateReport',
+            version='2021-03-20',
+            protocol='HTTPS',
+            pathname='/',
+            method='POST',
+            auth_type='AK',
+            style='RPC',
+            req_body_type='formData',
+            body_type='json'
+        )
+        return TeaCore.from_map(
+            emr_20210320_models.CreateReportResponse(),
+            self.call_api(params, req, runtime)
+        )
+
+    async def create_report_with_options_async(
+        self,
+        request: emr_20210320_models.CreateReportRequest,
+        runtime: util_models.RuntimeOptions,
+    ) -> emr_20210320_models.CreateReportResponse:
+        """
+        Currently we only support taihao platform
+        
+        @param request: CreateReportRequest
+        @param runtime: runtime options for this request RuntimeOptions
+        @return: CreateReportResponse
+        """
+        UtilClient.validate_model(request)
+        query = {}
+        if not UtilClient.is_unset(request.cluster_id):
+            query['ClusterId'] = request.cluster_id
+        if not UtilClient.is_unset(request.region_id):
+            query['RegionId'] = request.region_id
+        if not UtilClient.is_unset(request.select_timestamp):
+            query['SelectTimestamp'] = request.select_timestamp
+        req = open_api_models.OpenApiRequest(
+            query=OpenApiUtilClient.query(query)
+        )
+        params = open_api_models.Params(
+            action='CreateReport',
+            version='2021-03-20',
+            protocol='HTTPS',
+            pathname='/',
+            method='POST',
+            auth_type='AK',
+            style='RPC',
+            req_body_type='formData',
+            body_type='json'
+        )
+        return TeaCore.from_map(
+            emr_20210320_models.CreateReportResponse(),
+            await self.call_api_async(params, req, runtime)
+        )
+
+    def create_report(
+        self,
+        request: emr_20210320_models.CreateReportRequest,
+    ) -> emr_20210320_models.CreateReportResponse:
+        """
+        Currently we only support taihao platform
+        
+        @param request: CreateReportRequest
+        @return: CreateReportResponse
+        """
+        runtime = util_models.RuntimeOptions()
+        return self.create_report_with_options(request, runtime)
+
+    async def create_report_async(
+        self,
+        request: emr_20210320_models.CreateReportRequest,
+    ) -> emr_20210320_models.CreateReportResponse:
+        """
+        Currently we only support taihao platform
+        
+        @param request: CreateReportRequest
+        @return: CreateReportResponse
+        """
+        runtime = util_models.RuntimeOptions()
+        return await self.create_report_with_options_async(request, runtime)
+
     def decrease_nodes_with_options(
         self,
         request: emr_20210320_models.DecreaseNodesRequest,
@@ -947,6 +1051,106 @@ class Client(OpenApiClient):
     ) -> emr_20210320_models.GetClusterResponse:
         runtime = util_models.RuntimeOptions()
         return await self.get_cluster_with_options_async(request, runtime)
+
+    def get_cluster_clone_meta_with_options(
+        self,
+        request: emr_20210320_models.GetClusterCloneMetaRequest,
+        runtime: util_models.RuntimeOptions,
+    ) -> emr_20210320_models.GetClusterCloneMetaResponse:
+        """
+        获取集群克隆详情。
+        
+        @param request: GetClusterCloneMetaRequest
+        @param runtime: runtime options for this request RuntimeOptions
+        @return: GetClusterCloneMetaResponse
+        """
+        UtilClient.validate_model(request)
+        query = {}
+        if not UtilClient.is_unset(request.cluster_id):
+            query['ClusterId'] = request.cluster_id
+        if not UtilClient.is_unset(request.region_id):
+            query['RegionId'] = request.region_id
+        req = open_api_models.OpenApiRequest(
+            query=OpenApiUtilClient.query(query)
+        )
+        params = open_api_models.Params(
+            action='GetClusterCloneMeta',
+            version='2021-03-20',
+            protocol='HTTPS',
+            pathname='/',
+            method='POST',
+            auth_type='AK',
+            style='RPC',
+            req_body_type='formData',
+            body_type='json'
+        )
+        return TeaCore.from_map(
+            emr_20210320_models.GetClusterCloneMetaResponse(),
+            self.call_api(params, req, runtime)
+        )
+
+    async def get_cluster_clone_meta_with_options_async(
+        self,
+        request: emr_20210320_models.GetClusterCloneMetaRequest,
+        runtime: util_models.RuntimeOptions,
+    ) -> emr_20210320_models.GetClusterCloneMetaResponse:
+        """
+        获取集群克隆详情。
+        
+        @param request: GetClusterCloneMetaRequest
+        @param runtime: runtime options for this request RuntimeOptions
+        @return: GetClusterCloneMetaResponse
+        """
+        UtilClient.validate_model(request)
+        query = {}
+        if not UtilClient.is_unset(request.cluster_id):
+            query['ClusterId'] = request.cluster_id
+        if not UtilClient.is_unset(request.region_id):
+            query['RegionId'] = request.region_id
+        req = open_api_models.OpenApiRequest(
+            query=OpenApiUtilClient.query(query)
+        )
+        params = open_api_models.Params(
+            action='GetClusterCloneMeta',
+            version='2021-03-20',
+            protocol='HTTPS',
+            pathname='/',
+            method='POST',
+            auth_type='AK',
+            style='RPC',
+            req_body_type='formData',
+            body_type='json'
+        )
+        return TeaCore.from_map(
+            emr_20210320_models.GetClusterCloneMetaResponse(),
+            await self.call_api_async(params, req, runtime)
+        )
+
+    def get_cluster_clone_meta(
+        self,
+        request: emr_20210320_models.GetClusterCloneMetaRequest,
+    ) -> emr_20210320_models.GetClusterCloneMetaResponse:
+        """
+        获取集群克隆详情。
+        
+        @param request: GetClusterCloneMetaRequest
+        @return: GetClusterCloneMetaResponse
+        """
+        runtime = util_models.RuntimeOptions()
+        return self.get_cluster_clone_meta_with_options(request, runtime)
+
+    async def get_cluster_clone_meta_async(
+        self,
+        request: emr_20210320_models.GetClusterCloneMetaRequest,
+    ) -> emr_20210320_models.GetClusterCloneMetaResponse:
+        """
+        获取集群克隆详情。
+        
+        @param request: GetClusterCloneMetaRequest
+        @return: GetClusterCloneMetaResponse
+        """
+        runtime = util_models.RuntimeOptions()
+        return await self.get_cluster_clone_meta_with_options_async(request, runtime)
 
     def get_doctor_application_with_options(
         self,
@@ -2661,13 +2865,6 @@ class Client(OpenApiClient):
         request: emr_20210320_models.IncreaseNodesRequest,
         runtime: util_models.RuntimeOptions,
     ) -> emr_20210320_models.IncreaseNodesResponse:
-        """
-        扩容节点。
-        
-        @param request: IncreaseNodesRequest
-        @param runtime: runtime options for this request RuntimeOptions
-        @return: IncreaseNodesResponse
-        """
         UtilClient.validate_model(request)
         query = {}
         if not UtilClient.is_unset(request.application_configs):
@@ -2678,6 +2875,8 @@ class Client(OpenApiClient):
             query['ClusterId'] = request.cluster_id
         if not UtilClient.is_unset(request.increase_node_count):
             query['IncreaseNodeCount'] = request.increase_node_count
+        if not UtilClient.is_unset(request.min_increase_node_count):
+            query['MinIncreaseNodeCount'] = request.min_increase_node_count
         if not UtilClient.is_unset(request.node_group_id):
             query['NodeGroupId'] = request.node_group_id
         if not UtilClient.is_unset(request.payment_duration):
@@ -2710,13 +2909,6 @@ class Client(OpenApiClient):
         request: emr_20210320_models.IncreaseNodesRequest,
         runtime: util_models.RuntimeOptions,
     ) -> emr_20210320_models.IncreaseNodesResponse:
-        """
-        扩容节点。
-        
-        @param request: IncreaseNodesRequest
-        @param runtime: runtime options for this request RuntimeOptions
-        @return: IncreaseNodesResponse
-        """
         UtilClient.validate_model(request)
         query = {}
         if not UtilClient.is_unset(request.application_configs):
@@ -2727,6 +2919,8 @@ class Client(OpenApiClient):
             query['ClusterId'] = request.cluster_id
         if not UtilClient.is_unset(request.increase_node_count):
             query['IncreaseNodeCount'] = request.increase_node_count
+        if not UtilClient.is_unset(request.min_increase_node_count):
+            query['MinIncreaseNodeCount'] = request.min_increase_node_count
         if not UtilClient.is_unset(request.node_group_id):
             query['NodeGroupId'] = request.node_group_id
         if not UtilClient.is_unset(request.payment_duration):
@@ -2758,12 +2952,6 @@ class Client(OpenApiClient):
         self,
         request: emr_20210320_models.IncreaseNodesRequest,
     ) -> emr_20210320_models.IncreaseNodesResponse:
-        """
-        扩容节点。
-        
-        @param request: IncreaseNodesRequest
-        @return: IncreaseNodesResponse
-        """
         runtime = util_models.RuntimeOptions()
         return self.increase_nodes_with_options(request, runtime)
 
@@ -2771,12 +2959,6 @@ class Client(OpenApiClient):
         self,
         request: emr_20210320_models.IncreaseNodesRequest,
     ) -> emr_20210320_models.IncreaseNodesResponse:
-        """
-        扩容节点。
-        
-        @param request: IncreaseNodesRequest
-        @return: IncreaseNodesResponse
-        """
         runtime = util_models.RuntimeOptions()
         return await self.increase_nodes_with_options_async(request, runtime)